+++ CHANGELOG.md
@@ -1,13 +1,16 @@
 # python-utils
-- auto_utils
-- auto_mongo
-- auto_fm
+- common.utils
+- common.mongo
+- common.fm
 
 
 <br><br>
 
 ## auto_utils.py
 ---
+
+### 1.21.0
+- add change_to to config.file_type()
 
 ### 1.18
 - add `ts()`, returns ISO formatted UTC (now) -- non-ISO date, str-based or actual object
@@ -64,7 +67,7 @@
 ---
 
 ### 1.3.5
-- add with_sync_id feature to update() to a randomly generated _sync_id to docs 
+- add with_sync_id feature to update() to a randomly generated _sync_id to docs
 
 ### 1.3.1
 - 2nd update, keeping same version: bugfix with db_name when using `db_config`
@@ -107,7 +110,21 @@
 ## auto_fm.py
 ---
 
-### 2.4
+### 2.5.0
+- updated strings to f-string
+- added fm.fullpath, fm.cd(<relative-path>), fm.pwd()
+- add support to fm.ls() to return file/dir object w/ better details
+- add support to create/delete subfolders
+- added auto_create feature to dir_struct() and set_bucket()
+
+### 2.4.2
+- remove conditional general path update
+- bugfix: fm.ls() returning fn_only (filname only)
+
+### 2.4.1
+- ???
+
+### 2.4.0
 - change default directory generation to ['archive', 'errored', 'input', 'output']
 - add flag fn_only to ls(), latest(), oldest()
 - optimize dir_struct()